feat: configurable server address
