+++ config/default.json
@@ -1,4 +1,7 @@
 {
+  "server": {
+    "address": "0.0.0.0:9999"
+  },
   "acropolis": {
     "global": {
       "startup": {
